+++ examples/strategies/betfair_test_strategy.py
@@ -18,7 +18,7 @@ from decimal import Decimal
 from nautilus_trader.adapters.betfair.data import InstrumentSearch
 from nautilus_trader.common.enums import LogColor
 from nautilus_trader.core.message import Event
-from nautilus_trader.model.c_enums.orderbook_level import OrderBookLevel
+from nautilus_trader.model.c_enums.book_level import BookLevel
 from nautilus_trader.model.c_enums.time_in_force import TimeInForce
 from nautilus_trader.model.data import Data
 from nautilus_trader.model.data import DataType
@@ -99,11 +99,11 @@ class BetfairTestStrategy(TradingStrategy):
             # Subscribe to live data
             self.subscribe_order_book_deltas(
                 instrument_id=self.instrument_id,
-                level=OrderBookLevel.L2,
+                level=BookLevel.L2,
             )
             self.book = OrderBook(
                 instrument_id=self.instrument_id,
-                level=OrderBookLevel.L2,
+                level=BookLevel.L2,
             )
 
     def on_order_book_delta(self, delta: OrderBookDelta):