+++ nautilus_trader/model/orderbook/book.pyx
@@ -13,6 +13,7 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
+import json
 from operator import itemgetter
 
 import pandas as pd
@@ -20,12 +21,12 @@ from tabulate import tabulate
 
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.model.c_enums.aggressor_side cimport AggressorSide
+from nautilus_trader.model.c_enums.book_level cimport BookLevel
+from nautilus_trader.model.c_enums.book_level cimport BookLevelParser
 from nautilus_trader.model.c_enums.delta_type cimport DeltaType
 from nautilus_trader.model.c_enums.delta_type cimport DeltaTypeParser
 from nautilus_trader.model.c_enums.order_side cimport OrderSide
 from nautilus_trader.model.c_enums.order_side cimport OrderSideParser
-from nautilus_trader.model.c_enums.orderbook_level cimport OrderBookLevel
-from nautilus_trader.model.c_enums.orderbook_level cimport OrderBookLevelParser
 from nautilus_trader.model.data cimport Data
 from nautilus_trader.model.instruments.base cimport Instrument
 from nautilus_trader.model.orderbook.ladder cimport Ladder
@@ -47,7 +48,7 @@ cdef class OrderBook:
     def __init__(
         self,
         InstrumentId instrument_id not None,
-        OrderBookLevel level,
+        BookLevel level,
         int price_precision,
         int size_precision,
     ):
@@ -58,7 +59,7 @@ cdef class OrderBook:
         ----------
         instrument_id : InstrumentId
             The instrument identifier for the book.
-        level : OrderBookLevel
+        level : BookLevel
             The order book level (L1, L2, L3).
         price_precision : int
             The price precision for the book.
@@ -99,7 +100,7 @@ cdef class OrderBook:
     @staticmethod
     def create(
         Instrument instrument,
-        OrderBookLevel level,
+        BookLevel level,
     ):
         """
         Create a new order book with the given parameters.
@@ -108,7 +109,7 @@ cdef class OrderBook:
         ----------
         instrument : Instrument
             The instrument for the book.
-        level : OrderBookLevel
+        level : BookLevel
             The order book level (L1, L2, L3).
 
         Returns
@@ -119,19 +120,19 @@ cdef class OrderBook:
         Condition.not_none(instrument, "instrument")
         Condition.in_range_int(level, 1, 3, "level")
 
-        if level == OrderBookLevel.L1:
+        if level == BookLevel.L1:
             return L1OrderBook(
                 instrument_id=instrument.id,
                 price_precision=instrument.price_precision,
                 size_precision=instrument.size_precision,
             )
-        elif level == OrderBookLevel.L2:
+        elif level == BookLevel.L2:
             return L2OrderBook(
                 instrument_id=instrument.id,
                 price_precision=instrument.price_precision,
                 size_precision=instrument.size_precision,
             )
-        elif level == OrderBookLevel.L3:
+        elif level == BookLevel.L3:
             return L3OrderBook(
                 instrument_id=instrument.id,
                 price_precision=instrument.price_precision,
@@ -247,14 +248,14 @@ cdef class OrderBook:
         for bid in snapshot.bids:
             order = Order(
                 price=bid[0],
-                volume=bid[1],
+                size=bid[1],
                 side=OrderSide.BUY
             )
             self.update(order=order)
         for ask in snapshot.asks:
             order = Order(
                 price=ask[0],
-                volume=ask[1],
+                size=ask[1],
                 side=OrderSide.SELL
             )
             self.update(order=order)
@@ -518,7 +519,7 @@ cdef class OrderBook:
         else:
             return None
 
-    cpdef str pprint(self, int num_levels=3, show="volume"):
+    cpdef str pprint(self, int num_levels=3, show="size"):
         """
         Print the order book in a clear format.
 
@@ -719,7 +720,7 @@ cdef class L3OrderBook(OrderBook):
         """
         super().__init__(
             instrument_id=instrument_id,
-            level=OrderBookLevel.L3,
+            level=BookLevel.L3,
             price_precision=price_precision,
             size_precision=size_precision,
         )
@@ -753,7 +754,7 @@ cdef class L2OrderBook(OrderBook):
         """
         super().__init__(
             instrument_id=instrument_id,
-            level=OrderBookLevel.L2,
+            level=BookLevel.L2,
             price_precision=price_precision,
             size_precision=size_precision,
         )
@@ -865,7 +866,7 @@ cdef class L1OrderBook(OrderBook):
         """
         super().__init__(
             instrument_id=instrument_id,
-            level=OrderBookLevel.L1,
+            level=BookLevel.L1,
             price_precision=price_precision,
             size_precision=size_precision,
         )
@@ -952,7 +953,7 @@ cdef class L1OrderBook(OrderBook):
         else:
             self._top_bid_level.price = price
             self._top_bid.update_price(price)
-            self._top_bid.update_volume(size)
+            self._top_bid.update_size(size)
 
     cdef void _update_ask(self, double price, double size):
         if self._top_ask is None:
@@ -963,7 +964,7 @@ cdef class L1OrderBook(OrderBook):
         else:
             self._top_ask_level.price = price
             self._top_ask.update_price(price)
-            self._top_ask.update_volume(size)
+            self._top_ask.update_size(size)
 
     cpdef void delete(self, Order order) except *:
         """
@@ -1013,7 +1014,7 @@ cdef class OrderBookData(Data):
     def __init__(
         self,
         InstrumentId instrument_id not None,
-        OrderBookLevel level,
+        BookLevel level,
         uint64_t ts_event_ns,
         uint64_t ts_recv_ns,
     ):
@@ -1024,7 +1025,7 @@ cdef class OrderBookData(Data):
         ----------
         instrument_id : InstrumentId
             The instrument identifier for the book.
-        level : OrderBookLevel
+        level : BookLevel
             The order book level (L1, L2, L3).
         ts_event_ns: uint64
             The UNIX timestamp (nanoseconds) when data event occurred.
@@ -1046,7 +1047,7 @@ cdef class OrderBookSnapshot(OrderBookData):
     def __init__(
         self,
         InstrumentId instrument_id not None,
-        OrderBookLevel level,
+        BookLevel level,
         list bids not None,
         list asks not None,
         uint64_t ts_event_ns,
@@ -1059,7 +1060,7 @@ cdef class OrderBookSnapshot(OrderBookData):
         ----------
         instrument_id : InstrumentId
             The instrument identifier for the book.
-        level : OrderBookLevel
+        level : BookLevel
             The order book level (L1, L2, L3).
         bids : list
             The bids for the snapshot.
@@ -1076,13 +1077,66 @@ cdef class OrderBookSnapshot(OrderBookData):
         self.bids = bids
         self.asks = asks
 
+    def __eq__(self, OrderBookSnapshot other) -> bool:
+        return self.to_dict() == other.to_dict()
+
+    def __hash__(self) -> int:
+        return hash(frozenset(self.to_dict()))
+
     def __repr__(self) -> str:
         return (f"{type(self).__name__}("
                 f"'{self.instrument_id}', "
-                f"level={OrderBookLevelParser.to_str(self.level)}, "
+                f"level={BookLevelParser.to_str(self.level)}, "
                 f"bids={self.bids}, "
                 f"asks={self.asks}, "
                 f"ts_recv_ns={self.ts_recv_ns})")
+
+    @staticmethod
+    cdef OrderBookSnapshot from_dict_c(dict values):
+        return OrderBookSnapshot(
+            instrument_id=InstrumentId.from_str_c(values["instrument_id"]),
+            level=BookLevelParser.from_str(values["level"]),
+            bids=json.loads(values["bids"]),
+            asks=json.loads(values["asks"]),
+            ts_event_ns=values["ts_event_ns"],
+            ts_recv_ns=values["ts_recv_ns"],
+        )
+
+    @staticmethod
+    def from_dict(dict values):
+        """
+        Return an order book snapshot from the given dict values.
+
+        Parameters
+        ----------
+        values : dict[str, object]
+            The values for initialization.
+
+        Returns
+        -------
+        OrderBookSnapshot
+
+        """
+        return OrderBookSnapshot.from_dict_c(values)
+
+    cpdef dict to_dict(self):
+        """
+        Return a dictionary representation of this object.
+
+        Returns
+        -------
+        dict[str, object]
+
+        """
+        return {
+            "type": type(self).__name__,
+            "instrument_id": self.instrument_id.value,
+            "level": BookLevelParser.to_str(self.level),
+            "bids": json.dumps(self.bids),
+            "asks": json.dumps(self.asks),
+            "ts_event_ns": self.ts_event_ns,
+            "ts_recv_ns": self.ts_recv_ns,
+        }
 
 
 cdef class OrderBookDeltas(OrderBookData):
@@ -1093,7 +1147,7 @@ cdef class OrderBookDeltas(OrderBookData):
     def __init__(
         self,
         InstrumentId instrument_id not None,
-        OrderBookLevel level,
+        BookLevel level,
         list deltas not None,
         uint64_t ts_event_ns,
         uint64_t ts_recv_ns,
@@ -1105,7 +1159,7 @@ cdef class OrderBookDeltas(OrderBookData):
         ----------
         instrument_id : InstrumentId
             The instrument identifier for the book.
-        level : OrderBookLevel
+        level : BookLevel
             The order book level (L1, L2, L3).
         deltas : list[OrderBookDelta]
             The list of order book changes.
@@ -1119,12 +1173,63 @@ cdef class OrderBookDeltas(OrderBookData):
 
         self.deltas = deltas
 
+    def __eq__(self, OrderBookDeltas other) -> bool:
+        return self.to_dict() == other.to_dict()
+
+    def __hash__(self) -> int:
+        return hash(frozenset(self.to_dict()))
+
     def __repr__(self) -> str:
         return (f"{type(self).__name__}("
                 f"'{self.instrument_id}', "
-                f"level={OrderBookLevelParser.to_str(self.level)}, "
+                f"level={BookLevelParser.to_str(self.level)}, "
                 f"{self.deltas}, "
                 f"ts_recv_ns={self.ts_recv_ns})")
+
+    @staticmethod
+    cdef OrderBookDeltas from_dict_c(dict values):
+        return OrderBookDeltas(
+            instrument_id=InstrumentId.from_str_c(values["instrument_id"]),
+            level=BookLevelParser.from_str(values["level"]),
+            deltas=[OrderBookDelta.from_dict_c(d) for d in json.loads(values["deltas"])],
+            ts_event_ns=values["ts_event_ns"],
+            ts_recv_ns=values["ts_recv_ns"],
+        )
+
+    @staticmethod
+    def from_dict(dict values):
+        """
+        Return order book deltas from the given dict values.
+
+        Parameters
+        ----------
+        values : dict[str, object]
+            The values for initialization.
+
+        Returns
+        -------
+        OrderBookDeltas
+
+        """
+        return OrderBookDeltas.from_dict_c(values)
+
+    cpdef dict to_dict(self):
+        """
+        Return a dictionary representation of this object.
+
+        Returns
+        -------
+        dict[str, object]
+
+        """
+        return {
+            "type": type(self).__name__,
+            "instrument_id": self.instrument_id.value,
+            "level": BookLevelParser.to_str(self.level),
+            "deltas": json.dumps([d.to_dict() for d in self.deltas]),
+            "ts_event_ns": self.ts_event_ns,
+            "ts_recv_ns": self.ts_recv_ns,
+        }
 
 
 cdef class OrderBookDelta(OrderBookData):
@@ -1134,8 +1239,8 @@ cdef class OrderBookDelta(OrderBookData):
 
     def __init__(
         self,
-        InstrumentId instrument_id,
-        OrderBookLevel level,
+        InstrumentId instrument_id not None,
+        BookLevel level,
         DeltaType delta_type,
         Order order not None,
         uint64_t ts_event_ns,
@@ -1146,6 +1251,10 @@ cdef class OrderBookDelta(OrderBookData):
 
         Parameters
         ----------
+        instrument_id : InstrumentId
+            The instrument identifier.
+        level : BookLevel
+            The book level for the delta.
         delta_type : DeltaType
             The type of change (ADD, UPDATED, DELETE, CLEAR).
         order : Order
@@ -1161,10 +1270,63 @@ cdef class OrderBookDelta(OrderBookData):
         self.type = delta_type
         self.order = order
 
+    def __eq__(self, OrderBookDelta other) -> bool:
+        return self.to_dict() == other.to_dict()
+
+    def __hash__(self) -> int:
+        return hash(frozenset(self.to_dict()))
+
     def __repr__(self) -> str:
         return (f"{type(self).__name__}("
                 f"'{self.instrument_id}', "
-                f"level={OrderBookLevelParser.to_str(self.level)}, "
+                f"level={BookLevelParser.to_str(self.level)}, "
                 f"delta_type={DeltaTypeParser.to_str(self.type)}, "
                 f"order={self.order}, "
                 f"ts_recv_ns={self.ts_recv_ns})")
+
+    @staticmethod
+    cdef OrderBookDelta from_dict_c(dict values):
+        return OrderBookDelta(
+            instrument_id=InstrumentId.from_str_c(values["instrument_id"]),
+            level=BookLevelParser.from_str(values["level"]),
+            delta_type=DeltaTypeParser.from_str(values["delta_type"]),
+            order=Order.from_dict_c(json.loads(values["order"])),
+            ts_event_ns=values["ts_event_ns"],
+            ts_recv_ns=values["ts_recv_ns"],
+        )
+
+    @staticmethod
+    def from_dict(dict values):
+        """
+        Return an order book delta from the given dict values.
+
+        Parameters
+        ----------
+        values : dict[str, object]
+            The values for initialization.
+
+        Returns
+        -------
+        OrderBookDelta
+
+        """
+        return OrderBookDelta.from_dict_c(values)
+
+    cpdef dict to_dict(self):
+        """
+        Return a dictionary representation of this object.
+
+        Returns
+        -------
+        dict[str, object]
+
+        """
+        return {
+            "type": type(self).__name__,
+            "instrument_id": self.instrument_id.value,
+            "level": BookLevelParser.to_str(self.level),
+            "delta_type": DeltaTypeParser.to_str(self.type),
+            "order": json.dumps(self.order.to_dict()),
+            "ts_event_ns": self.ts_event_ns,
+            "ts_recv_ns": self.ts_recv_ns,
+        }