+++ tests/unit_tests/data/test_data_engine.py
@@ -33,7 +33,7 @@ from nautilus_trader.model.data import Data
 from nautilus_trader.model.data import DataType
 from nautilus_trader.model.enums import AggressorSide
 from nautilus_trader.model.enums import BarAggregation
-from nautilus_trader.model.enums import OrderBookLevel
+from nautilus_trader.model.enums import BookLevel
 from nautilus_trader.model.enums import PriceType
 from nautilus_trader.model.identifiers import ClientId
 from nautilus_trader.model.identifiers import InstrumentId
@@ -921,7 +921,7 @@ class DataEngineTests(unittest.TestCase):
                 OrderBook,
                 {
                     "InstrumentId": ETHUSDT_BINANCE.id,
-                    "Level": OrderBookLevel.L2,
+                    "Level": BookLevel.L2,
                     "Depth": 25,
                     "Interval": 0,  # Streaming
                 },
@@ -935,7 +935,7 @@ class DataEngineTests(unittest.TestCase):
 
         snapshot = OrderBookSnapshot(
             instrument_id=ETHUSDT_BINANCE.id,
-            level=OrderBookLevel.L2,
+            level=BookLevel.L2,
             bids=[[1000, 1]],
             asks=[[1001, 1]],
             ts_event_ns=0,
@@ -966,7 +966,7 @@ class DataEngineTests(unittest.TestCase):
                 OrderBook,
                 {
                     "InstrumentId": ETHUSDT_BINANCE.id,
-                    "Level": OrderBookLevel.L2,
+                    "Level": BookLevel.L2,
                     "Depth": 25,
                     "Interval": 0,  # Streaming
                 },
@@ -980,7 +980,7 @@ class DataEngineTests(unittest.TestCase):
 
         deltas = OrderBookDeltas(
             instrument_id=ETHUSDT_BINANCE.id,
-            level=OrderBookLevel.L2,
+            level=BookLevel.L2,
             deltas=[],
             ts_event_ns=0,
             ts_recv_ns=0,
@@ -1012,7 +1012,7 @@ class DataEngineTests(unittest.TestCase):
                 OrderBook,
                 {
                     "InstrumentId": ETHUSDT_BINANCE.id,
-                    "Level": OrderBookLevel.L2,
+                    "Level": BookLevel.L2,
                     "Depth": 25,
                     "Interval": 0,  # Streaming
                 },
@@ -1029,7 +1029,7 @@ class DataEngineTests(unittest.TestCase):
                 OrderBook,
                 {
                     "InstrumentId": ETHUSDT_BINANCE.id,
-                    "Level": OrderBookLevel.L2,
+                    "Level": BookLevel.L2,
                     "Depth": 25,
                     "Interval": 0,  # Streaming
                 },
@@ -1044,7 +1044,7 @@ class DataEngineTests(unittest.TestCase):
 
         snapshot = OrderBookSnapshot(
             instrument_id=ETHUSDT_BINANCE.id,
-            level=OrderBookLevel.L2,
+            level=BookLevel.L2,
             bids=[[1000, 1]],
             asks=[[1001, 1]],
             ts_event_ns=0,